+++ srcs/core/commands/CmdMode.cpp
@@ -1,0 +1,6 @@
+#include "Server.hpp"
+
+void Server::handleMode(int fd, std::istringstream& iss)
+{
+	(void)fd; std::cout<<iss;
+}
