+++ srcs/core/commands/CmdTopic.cpp
@@ -1,0 +1,6 @@
+#include "Server.hpp"
+
+void Server::handleTopic(int fd, std::istringstream& iss)
+{
+	(void)fd; std::cout<<iss;
+}
